Migrate to flutter 3.20.0.pre (#57)
Migrate to `3.20.0-1.1.pre`. This is necessary because the current SDK used to analyze packages on pub.dev is `3.20.0-1.1.pre`, which has a breaking change in the `PageView` class that is not compatible with the current version of this package (described below). As a result, static analysis errors occur on pub.dev.

## What changes will be made in 3.20?

`PageView.controller` will be nullable.

## How to migrate?

Make `OverflowPageView` always receive a non-nullable `PageController` in the constructor.